Event Organiser: introduce events navigation
Considering event archives as adjacent archives, the posts navigation for
previous/next events new reflects this relationship. So yearly archives
link to previous/next years, as do months and days. The links are checked
for existing events in the respective archives, so only valid links are
displayed. We're also accounting for an archive's pagination for previous
archives as they're entered from the end of the post query.